feat: scrolling window
