+++ src/components/Session/SessionInfo.js
@@ -4,11 +4,8 @@ import { Star } from "react-feather";
 import "./Session.scss";
 export default function SessionInfo({ value }) {
   return (
-    <div className="session-info">
-      <Box h={10} className="sticky-header"></Box>
-      <div>
-        <Image src={value.image} />
-      </div>
-    </div>
+    <Box>
+      <Image src={value.image} />
+    </Box>
   );
 }
